change: Move OCI fixture closer to the usage
